+++ src/main/java/pvpmode/compatibility/modules/lotr/LOTRModCompatibilityModule.java
@@ -1,6 +1,7 @@
 package pvpmode.compatibility.modules.lotr;
 
 import java.io.*;
+import java.lang.reflect.Field;
 import java.nio.file.*;
 import java.util.*;
 
@@ -11,7 +12,8 @@ import lotr.common.entity.npc.LOTREntityNPC;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.*;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.entity.living.LivingSetAttackTargetEvent;
+import net.minecraftforge.event.entity.living.*;
+import net.minecraftforge.event.entity.player.*;
 import pvpmode.*;
 import pvpmode.compatibility.CompatibilityModule;
 import pvpmode.compatibility.events.*;
@@ -31,6 +33,9 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
 
     private boolean areEnemyBiomeOverridesEnabled;
     private boolean blockFTInPvP;
+    private boolean dropSkullWithKeepInventory;
+
+    private LOTREventHandler eventHandler;
 
     @Override
     public void load () throws IOException
@@ -42,6 +47,9 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
             "If true, the PvP mode enemy biome override condition for LOTR biomes will be enabled. Players who are an enemy of a faction are forced to have PvP enabled while they're in a biome which is clearly assignable to that faction. This is highly configurable.");
         blockFTInPvP = PvPMode.config.getBoolean ("Block fast traveling in PvP", LOTR_CONFIGURATION_CATEGORY,
             true, "If enabled, players cannot use the LOTR fast travel system while they're in PvP.");
+        dropSkullWithKeepInventory = PvPMode.config.getBoolean ("Always Drop Player Skulls",
+            LOTR_CONFIGURATION_CATEGORY, true,
+            "If true, players killed with a weapon with the headhunting modifier will drop their skulls even with keepInventory enabled.");
 
         PvPMode.config.addCustomCategoryComment (LOTR_CONFIGURATION_CATEGORY,
             "Configuration entries for compatibility with the \"The Lord of the Rings Minecraft Mod\"");
@@ -54,6 +62,8 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
         {
             initEnemyBiomeOverrides (configurationFolder);
         }
+
+        this.retrieveLOTREventHandler ();
     }
 
     private void initEnemyBiomeOverrides (Path configurationFolder) throws IOException
@@ -225,6 +235,28 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
         FMLLog.info ("Recreated the LOTR biome id file");
     }
 
+    private void retrieveLOTREventHandler ()
+    {
+        try
+        {
+            Field eventHandlerField = LOTRMod.class.getDeclaredField ("modEventHandler");
+            eventHandlerField.setAccessible (true);
+            this.eventHandler = (LOTREventHandler) eventHandlerField.get (null);
+            if (this.eventHandler != null)
+            {
+                FMLLog.info ("Successfully retrieved the LOTR event handler");
+            }
+            else
+            {
+                FMLLog.warning ("Couldn't retrieve the LOTR event handler - features depending on it won't be enabled");
+            }
+        }
+        catch (Exception e)
+        {
+            FMLLog.getLogger ().error ("Couldn't retrieve the LOTR event handler", e);
+        }
+    }
+
     @SubscribeEvent
     public void onEntityMasterExtraction (EntityMasterExtractionEvent event)
     {
@@ -287,6 +319,21 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
                         npc.setRevengeTarget (null);
                     }
                 }
+            }
+        }
+    }
+
+    @SubscribeEvent
+    public void onPlayerDeath (LivingDeathEvent event)
+    {
+        if (dropSkullWithKeepInventory && eventHandler != null
+            && event.entityLiving.worldObj.getGameRules ().getGameRuleBooleanValue ("keepInventory"))
+        {
+            if (event.entityLiving instanceof EntityPlayer)
+            {
+                EntityPlayer player = (EntityPlayer) event.entityLiving;
+                // The last two parameters are not used by the current implementation
+                this.eventHandler.onPlayerDrops (new PlayerDropsEvent (player, event.source, null, false));
             }
         }
     }